+++ libs/agent-sm/agent/src/main/java/org/opensearch/javaagent/RuntimeHaltInterceptor.java
@@ -17,20 +17,20 @@ import java.util.stream.Stream;
 import net.bytebuddy.asm.Advice;
 
 /**
- * {@link System#exit} interceptor
+ * {@link Runtime#halt} interceptor
  */
-public class SystemExitInterceptor {
+public class RuntimeHaltInterceptor {
     /**
-     * SystemExitInterceptor
+     * RuntimeHaltInterceptor
      */
-    public SystemExitInterceptor() {}
+    public RuntimeHaltInterceptor() {}
 
     /**
      * Interceptor
      * @param code exit code
      * @throws Exception exceptions
      */
-    @Advice.OnMethodEnter()
+    @Advice.OnMethodEnter
     @SuppressWarnings("removal")
     public static void intercept(int code) throws Exception {
         final Policy policy = AgentPolicy.getPolicy();
@@ -43,7 +43,7 @@ public class SystemExitInterceptor {
         final Stream<Class<?>> chain = walker.walk(StackCallerClassChainExtractor.INSTANCE);
 
         if (AgentPolicy.isChainThatCanExit(caller, chain) == false) {
-            throw new SecurityException("The class " + caller + " is not allowed to call System::exit(" + code + ")");
+            throw new SecurityException("The class " + caller + " is not allowed to call Runtime::halt(" + code + ")");
         }
     }
 }